# optimize integer binning